pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matrix-guardian"
version = "0.1.0"
description = "Matrix Guardian — AI control plane for the Matrix EcoSystem"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "psycopg[binary]>=3.1.0",
    "python-json-logger>=2.0.0",
    "tenacity>=8.3.0",
    "langgraph>=0.1.0",
    "pyyaml>=6.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]
